+++ filter-plugin/logstash-filter-teradatadb-guardium/src/main/java/com/ibm/guardium/teradatadb/Parser.java
@@ -21,7 +21,7 @@ public class Parser {
 		
 			record.setSessionId(e.getField(Constants.SESSION_ID).toString());
 			
-			record.setDbName(Constants.UNKNOWN_STRING);
+			record.setDbName(Constants.NOT_AVAILABLE);
 			
 			record.setAppUserName(Constants.UNKNOWN_STRING);
 			
@@ -66,7 +66,7 @@ public class Parser {
 			
 				sessionLocator.setClientIp(e.getField(Constants.CLIENT_IP).toString());
 				sessionLocator.setClientPort(Constants.DEFAULT_PORT);
-				sessionLocator.setServerIp(Constants.DEFAULT_IP);
+				sessionLocator.setServerIp(e.getField(Constants.SERVER_IP).toString());
 				sessionLocator.setServerPort(Constants.DEFAULT_PORT);
 				sessionLocator.setIpv6(false);
 				sessionLocator.setClientIpv6(Constants.UNKNOWN_STRING);
@@ -90,7 +90,7 @@ public class Parser {
 			accessor.setSourceProgram(e.getField(Constants.LOGON_SOURCE).toString());
 			accessor.setClient_mac(Constants.UNKNOWN_STRING);
 			accessor.setServerDescription(Constants.UNKNOWN_STRING);
-			accessor.setServiceName(Constants.UNKNOWN_STRING);
+			accessor.setServiceName(Constants.NOT_AVAILABLE);
 			accessor.setServerOs(Constants.UNKNOWN_STRING);
 			accessor.setServerHostName(e.getField(Constants.SERVER_HOSTNAME).toString());
 			accessor.setOsUser(e.getField(Constants.OS_USER).toString());