+++ input-plugin/logstash-input-mongo-atlas/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.7 - 01/05/2026
+- GRD-114688: Update input configuration to include mongo-api-url
+
+## 1.0.7
+- GRD-114688: Guardium Implementation for Mongo Atlas.
+
 ## 1.0.1
 - Fix gradle incompatibility issue with Logstash 7.9 and above.